@@ -1,126 +1,197 @@
-#ifndef GUARD_APPLICATION_H
-#define GUARD_APPLICATION_H
+﻿#pragma once
 
-#include <functional>
-#include <string>
+#include "PCH.h"
 
 struct GLFWwindow;
 
 /**
- * @brief ���ø����̼��� �����մϴ�.
+ * @brief 애플리케이션을 정의합니다.
  */
 class Application final
 {
-	using StartCallback = std::function<void()>;
-	using UpdateCallback = std::function<void(float)>;
-	using RenderCallback = std::function<void()>;
-	using CloseCallback = std::function<void()>;
 public:
-    /**
-    * @brief ���ø����̼� ������ �����մϴ�.
-    */
-    struct Specification final
-    {
-        /**
-         * @brief OpenGL ���ؽ�Ʈ ���� ����.
-         */
-        int majorVersion;
+	/**
+	 * @brief 애플리케이션 내 사용될 리소스를 불러오는 콜백을 정의합니다.
+	 */
+	using StartCallback = std::function<void()>;
 
-        /**
-         * @brief OpenGL ���ؽ�Ʈ ���� ����.
-         */
-        int minorVersion;
+	/**
+	 * @brief 애플리케이션이 매 프레임마다 업데이트되는 콜백을 정의합니다.
+	 */
+	using UpdateCallback = std::function<void(const float)>;
 
-        /**
-         * @brief ���ø����̼� Ÿ��Ʋ.
-         */
-        std::string name;
+	/**
+	 * @brief 애플리케이션이 매 프레임마다 그려지는 콜백을 정의합니다.
+	 */
+	using RenderCallback = std::function<void()>;
 
-        /**
-         * @brief ���ø����̼� �ʺ�.
-         */
-        int width;
+	/**
+	 * @brief 애플리케이션이 닫힐 때 호출되는 콜백을 정의합니다.
+	 */
+	using CloseCallback = std::function<void()>;
 
-        /**
-         * @brief ���ø����̼� ����.
-         */
-        int height;
-
-        /**
-         * @brief ���ø����̼��� ������Ʈ �ֱ�.
-         */
-        int fps;
-
-        /**
-         * @brief ���ø����̼��� ����� �� �����մϴ�.
-         */
-        StartCallback onStart;
-
-        /**
-         * @brief �� �����Ӹ��� ȣ��˴ϴ�.
-         */
-        UpdateCallback onUpdate;
-
-        /**
-         * @brief �������� �� ȣ��˴ϴ�.
-         */
-        RenderCallback onRender;
+	/**
+	 * @brief 애플리케이션 설정을 정의합니다.
+	 */
+	struct Configuration final
+	{
+		/**
+		 * @brief 창의 너비.
+		 */
+		int width;
 
 		/**
-		 * @brief ���ø����̼��� ����� �� ȣ��˴ϴ�.
+		 * @brief 창의 높이.
+		 */
+		int height;
+
+		/**
+		 * @brief 창의 제목.
+		 */
+		const char* title;
+
+		/**
+		 * @brief 전체 화면 모드 여부.
+		 */
+		bool shouldFullscreen;
+
+		/**
+		 * @brief 창 테두리 표시 여부.
+		 */
+		bool shouldDecorate;
+
+		/**
+		 * @brief 창 크기 조절 가능 여부.
+		 */
+		bool shouldResizable;
+
+		/**
+		 * @brief 수직 동기화(VSync) 사용 여부.
+		 */
+		bool shouldVSync;
+
+		/**
+		 * @brief 창이 켜질 때 호출되는 콜백 함수.
+		 */
+		StartCallback onStart;
+
+		/**
+		 * @brief 매 프레임마다 호출되는 콜백 함수.
+		 */
+		UpdateCallback onUpdate;
+
+		/**
+		 * @brief 창이 그려질 때 호출되는 콜백 함수.
+		 */
+		RenderCallback onRender;
+
+		/**
+		 * @brief 창이 닫힐 때 호출되는 콜백 함수.
 		 */
 		CloseCallback onClose;
-    };
-
-    /**
-     * @brief ���ø����̼� ������ ��ȯ�մϴ�.
-     * 
-	 * @return const Application::Specification& ���ø����̼� ����
-     */
-    [[nodiscard]]
-	static inline const Application::Specification& GetSpecification() noexcept;
+	};
 
 	/**
-	 * @brief ���ø����̼��� �����մϴ�.
-     * 
-	 * @param specification_ ���ø����̼� ����
+	 * @brief 애플리케이션 설정을 반환합니다.
+	 * 
+	 * @return Application::Configuration& 애플리케이션 설정
 	 */
-	static int Run(const Application::Specification& specification_) noexcept;
+	[[nodiscard]]
+	static inline constexpr Application::Configuration& GetSpecification() noexcept;
 
 	/**
-	 * @brief ���ø����̼��� �����մϴ�.
+	 * @brief 애플리케이션 설정을 지정합니다.
+	 * 
+	 * @param config_ 애플리케이션 설정
 	 */
-	static void Quit() noexcept;
+	static inline void SetSpecification(const Application::Configuration& config_) noexcept;
+
+	/**
+	 * @brief GLFW 윈도우 핸들를 반환합니다.
+	 * 
+	 * @return GLFWwindow* const GLFW 윈도우 핸들
+	 */
+	[[nodiscard]]
+	static inline constexpr GLFWwindow* const GetNativeWindow() noexcept;
+
+	/**
+	 * @brief GLFW 윈도우 핸들를 지정합니다.
+	 * 
+	 * @param window_ GLFW 윈도우 핸들
+	 */
+	static void SetNativeWindow(GLFWwindow* const window_) noexcept;
+
+	/**
+	 * @brief 애플리케이션을 실행합니다.
+	 * 
+	 * @param configuration_ 실행할 애플리케이션의 설정
+	 * 
+	 * @return int 종료 코드
+	 */
+	static int Run(const Application::Configuration& configuration_) noexcept;
+
+	/**
+	 * @brief 애플리케이션을 종료합니다.
+	 * 
+	 * @param exitCode_ 종료 코드
+	 */
+	static void Quit(const int exitCode_) noexcept;
 
 private:
 	/**
-	 * @brief �����츦 �ʱ�ȭ�մϴ�.
+	 * @brief 애플리케이션을 초기화합니다.
 	 * 
-	 * @return ���� ���θ� �ʱ�ȭ�մϴ�.
+	 * @return bool 초기화 성공 여부
 	 */
-	static bool InitWindow() noexcept;
+	static bool Initialize() noexcept;
 
 	/**
-	 * @brief �׷��Ƚ� API�� �ʱ�ȭ�մϴ�.
-	 * 
-	 * @return bool ���� ����
+	 * @brief 애플리케이션 내 리소스를 불러옵니다.
 	 */
-	static bool InitGraphics() noexcept;
+	static void Start() noexcept;
 
 	/**
-	 * @brief ���ø����̼� ����.
+	 * @brief 애플리케이션을 업데이트합니다.
 	 */
-	static Application::Specification specification;
+	static void Update(const float deltaTime_) noexcept;
 
 	/**
-	 * @brief ������ �ڵ�.
+	 * @brief 애플리케이션을 그립니다.
 	 */
+	static void Render() noexcept;
+
+	/**
+	 * @brief 애플리케이션 내 리소스를 정리하고, 애플리케이션을 닫습니다.
+	 */
+	static void Close() noexcept;
+
+	/**
+	 * @brief 애플리케이션 설정.
+	 */
+	static Application::Configuration configuration;
+
+	/**
+	* @brief GLFW 윈도우 핸들.
+	*/
 	static GLFWwindow* window;
 };
 
-inline const Application::Specification& Application::GetSpecification() noexcept
+inline constexpr Application::Configuration& Application::GetSpecification() noexcept
 {
-    return specification;
+	return configuration;
 }
 
-#endif // !GUARD_APPLICATION_H
+inline void Application::SetSpecification(const Application::Configuration& config_) noexcept
+{
+	configuration = config_;
+}
+
+inline constexpr GLFWwindow* const Application::GetNativeWindow() noexcept
+{
+	return window;
+}
+
+inline void Application::SetNativeWindow(GLFWwindow* const window_) noexcept
+{
+	window = window_;
+}
