@@ -1,231 +1,262 @@
-#ifndef GUARD_CAMERA_H
-#define GUARD_CAMERA_H
+﻿#pragma once
 
-#include <glm/vec3.hpp>
-#include <glm/ext.hpp>
+#include "PCH.h"
 
 #include "Application.h"
 #include "Shader.h"
 
 /**
- * @brief Scene �� ī�޶� �����մϴ�.
+ * @brief 
  */
 class Camera final
 {
 public:
-    /**
-     * @enum Projection
-     *
-     * @brief ī�޶��� ���� ����� �����մϴ�.
-     */
-    enum class Projection
-    {
-        /**
-         * @brief ���� ���� ���.
-         */
-        Orthographic,
+	/**
+	 * @brief 카메라의 투영 방식을 정의합니다.
+	 */
+	enum class Projection
+	{
+		/**
+		 * @brief 원근 투영.
+		 */
+		Perspective,
 
-        /**
-         * @brief ���� ���� ���.
-         */
-        Perspective
-    };
+		/**
+		 * @brief 직교 투영.
+		 */
+		Orthographic
+	};    
 
-    /**
-     * @brief ������.
-     *
-	 * @param projection_ ������ ī�޶��� ���� ���
-     * @param position_   ������ ī�޶��� ��ġ
-     * @param front_      ������ ī�޶� �ٶ󺸴� ����
-     * @param up_         ������ ī�޶��� ������ ����
-     */
-    explicit Camera(
-        const Camera::Projection projection_,
-        const glm::vec3&         position_,
-        const glm::vec3&         front_,
-        const glm::vec3&         up_
-    ) noexcept;
+	/**
+	 * @brief 뷰포트를 정의합니다.
+	 */
+	struct Viewport final
+	{
+		int x;
+		int y;
+		int width;
+		int height;
+	};
 
-    /**
-     * @brief �ش� ī�޶��� ��ġ�� �����ɴϴ�.
-     * 
-     * @return glm::vec3 �ش� ī�޶��� ��ġ
-     */
-    [[nodiscard]]
-    inline constexpr glm::vec3 GetPosition() const noexcept;
+	/**
+	 * @brief 생성자.
+	 */
+	explicit Camera() noexcept;
 
-    /**
-     * @brief �ش� ī�޶��� ��ġ�� ������ ��ġ�� �����մϴ�.
-     * 
-     * @param position_ ������ ��ġ
-     */
-    inline void SetPosition(const glm::vec3& position_) noexcept;
+	/**
+	 * @brief 해당 카메라의 투영 방식을 반환합니다.
+	 * 
+	 * @return Camera::Projection 해당 카메라의 투영 방식
+	 */
+	[[nodiscard]]
+	inline constexpr Camera::Projection GetProjection() const noexcept;
 
-    /**
-     * @brief �ش� ī�޶��� �չ��� ���͸� ��ȯ�մϴ�.
-     * 
-	 * @return glm::vec3 �ش� ī�޶��� �չ��� ����
-     */
-    [[nodiscard]]
-    inline constexpr glm::vec3 GetForward() const noexcept;
+	/**
+	 * @brief 해당 카메라의 투영 방식을 지정한 방식으로 설정합니다.
+	 * 
+	 * @param projection_ 지정할 투영 방식
+	 */
+	inline void SetProjection(Camera::Projection projection_) noexcept;
 
-    /**
-     * @brief �ش� ī�޶��� �չ��� ���͸� ������ ���ͷ� �����մϴ�.
-     * 
-     * @param position_ ������ �չ��� ����
-     */
-    inline void SetForward(const glm::vec3& position_) noexcept;
+	/**
+	 * @brief 해당 카메라의 위치를 반환합니다.
+	 *
+	 * @return glm::vec3 해당 카메라의 위치
+	 */
+	[[nodiscard]]
+	inline constexpr glm::vec3 GetPosition() const noexcept;
 
-    /**
-     * @brief �ش� ī�޶��� ������ ���͸� ��ȯ�մϴ�.
-     * 
-     * @return glm::vec3 �ش� ī�޶��� ������ ����
-     */
-    [[nodiscard]]
+	/**
+	 * @brief 해당 카메라의 위치를 지정한 위치로 설정합니다.
+	 *
+	 * @param position_ 지정할 위치
+	 */
+	inline void SetPosition(const glm::vec3& position_) noexcept;
+
+	/**
+	 * @brief 해당 카메라의 앞방향 벡터를 반환합니다.
+	 *
+	 * @return glm::vec3 해당 카메라의 앞방향 벡터
+	 */
+	[[nodiscard]]
+	inline constexpr glm::vec3 GetForward() const noexcept;
+
+	/**
+	 * @brief 해당 카메라의 앞방향 벡터를 지정한 벡터로 설정합니다.
+	 *
+	 * @param position_ 지정할 앞방향 벡터
+	 */
+	inline void SetForward(const glm::vec3& position_) noexcept;
+
+	/**
+	 * @brief 해당 카메라의 윗방향 벡터를 반환합니다.
+	 *
+	 * @return glm::vec3 해당 카메라의 윗방향 벡터
+	 */
+	[[nodiscard]]
 	inline constexpr glm::vec3 GetUp() const noexcept;
 
 	/**
-	 * @brief �ش� ī�޶��� ������ ���͸� ������ ���ͷ� �����մϴ�.
-     * 
-	 * @param up_ ������ ����.
+	 * @brief 해당 카메라의 윗방향 벡터를 지정한 벡터로 설정합니다.
+	 *
+	 * @param up_ 지정할 벡터.
 	 */
 	inline void SetUp(const glm::vec3& up_) noexcept;
 
-    /**
-     * @brief View ��� ��ȯ
-     */
-    [[nodiscard]]
-    inline glm::mat4 GetViewMatrix() const noexcept;
-
-    /**
-     * @brief Projection ��� ��ȯ (���� ��ü�� ����)
-     */
-    [[nodiscard]]
-    inline glm::mat4 GetProjectionMatrix() const noexcept;
-
-    /**
-     * @brief ��Ⱦ�� ������Ʈ (���� ��ü�� ����)
-     */
-    inline void SetAspectRatio(const float aspectRatio_) noexcept;
-
-    /**
-     * @brief ���� ����� ��Ÿ�ӿ� ��ü�մϴ�.
-     */
-    inline void SetProjection(Camera::Projection projection_) noexcept;
+	/**
+	 * @brief 뷰포트 반환
+	 */
+	[[nodiscard]]
+	inline constexpr Viewport GetViewport() noexcept;
 
 	/**
-	 * @brief 
-     * 
-	 * @param shader_ 
+	 * @brief 뷰포트 설정
+	 */
+	inline void SetViewport(const Viewport& viewport_) noexcept;
+
+	/**
+	 * @brief View 행렬 반환
+	 */
+	[[nodiscard]]
+	inline glm::mat4 GetViewMatrix() const noexcept;
+
+	/**
+	 * @brief Projection 행렬 반환
+	 */
+	[[nodiscard]]
+	inline glm::mat4 GetProjectionMatrix() const noexcept;
+
+	/**
+	 * @brief 렌더링 전에 호출되어 셰이더에 뷰 및 프로젝션 행렬을 설정합니다.
+	 * 
+	 * @param shader_ 셰이더 객체
 	 */
 	void PreRender(const Shader& shader_) const noexcept;
 
 private:
-    /**
-     * @brief �ش� ī�޶��� ���� ���.
-     */
-    Camera::Projection projection;
+	/**
+	* @brief 해당 카메라의 투영 방식.
+	*/
+	Camera::Projection projection = Camera::Projection::Perspective;
 
 	/**
-	 * @brief �ش� ī�޶��� ��ġ.
+	 * @brief 해당 카메라의 위치.
 	 */
-	glm::vec3 position;
+	glm::vec3 position = glm::vec3(0.0f);
 
 	/**
-	 * @brief �ش� ī�޶��� �չ��� ����.
+	 * @brief 해당 카메라의 앞방향 벡터.
 	 */
-	glm::vec3 forward;
+	glm::vec3 forward = glm::vec3(0.0f, 0.0f, 10.0f);
 
 	/**
-	 * @brief �ش� ī�޶��� ������ ����.
+	 * @brief 해당 카메라의 윗방향 벡터.
 	 */
-	glm::vec3 up;
+	glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f);
+	
+	/**
+	 * @brief 시야각.
+	 */
+	float fov = 45.0f;
 
-    /**
-     * @brief �þ߰�.
-     */
-    float fov = 45.0f;
+	/**
+	 * @brief 근평면.
+	 */
+	float nearPlane = 0.1f;
 
-    /**
-     * @brief 
-     */
-    float nearPlane = 0.1f;
+	/**
+	 * @brief 원평면.
+	 */
+	float farPlane = 100.0f;
 
-    /**
-     * @brief
-     */
-    float farPlane = 100.0f;
+	/**
+	 * @brief 
+	 */
+	float orthoSize = 10.0f;
 
-    /**
-     * @brief
-     */
-    float orthoSize = 10.0f;
+	/**
+	 * @brief 뷰포트.
+	 */
+	Viewport viewport;
 };
 
-inline constexpr glm::vec3 Camera::GetPosition() const noexcept
+inline constexpr Camera::Projection Camera::GetProjection() const noexcept
 {
-    return position;
-}
-
-inline void Camera::SetPosition(const glm::vec3& position_) noexcept
-{
-    position = position_;
-}
-
-inline constexpr glm::vec3 Camera::GetForward() const noexcept
-{
-    return forward;
-}
-
-inline void Camera::SetForward(const glm::vec3& forward_) noexcept
-{
-    forward = forward_;
-}
-
-inline constexpr glm::vec3 Camera::GetUp() const noexcept
-{
-    return up;
-}
-
-inline void Camera::SetUp(const glm::vec3& up_) noexcept
-{
-    up = up_;
-}
-
-inline glm::mat4 Camera::GetViewMatrix() const noexcept
-{
-    return glm::lookAt(position, position + forward, up);
-}
-
-inline glm::mat4 Camera::GetProjectionMatrix() const noexcept
-{
-    const float height      = static_cast<float>(Application::GetSpecification().height);
-    const float width       = static_cast<float>(Application::GetSpecification().width);
-    const float aspectRatio = width / height;
-
-    switch (projection)
-    {
-        case Projection::Orthographic:
-        {
-            const float halfHeight = orthoSize * 0.5f;
-            const float halfWidth  = halfHeight * aspectRatio;
-
-            return glm::ortho(-halfWidth, halfWidth, -halfHeight, halfHeight, nearPlane, farPlane);
-        }
-        case Projection::Perspective:
-        {
-            return glm::perspective(glm::radians(fov), aspectRatio, nearPlane, farPlane);
-        }
-        default:
-        {
-            return { 1.0f };
-        }
-    }
+	return projection;
 }
 
 inline void Camera::SetProjection(Camera::Projection projection_) noexcept
 {
-    projection = projection_;
+	projection = projection_;
 }
 
-#endif // !GUARD_CAMERA_H
+inline constexpr glm::vec3 Camera::GetPosition() const noexcept
+{
+	return position;
+}
+
+inline void Camera::SetPosition(const glm::vec3& position_) noexcept
+{
+	position = position_;
+}
+
+inline constexpr glm::vec3 Camera::GetForward() const noexcept
+{
+	return forward;
+}
+
+inline void Camera::SetForward(const glm::vec3& forward_) noexcept
+{
+	forward = forward_;
+}
+
+inline constexpr glm::vec3 Camera::GetUp() const noexcept
+{
+	return up;
+}
+
+inline void Camera::SetUp(const glm::vec3& up_) noexcept
+{
+	up = up_;
+}
+
+inline constexpr Camera::Viewport Camera::GetViewport() noexcept
+{
+	return viewport;
+}
+
+inline void Camera::SetViewport(const Viewport& viewport_) noexcept
+{
+	viewport = viewport_;
+}
+
+inline glm::mat4 Camera::GetViewMatrix() const noexcept
+{
+	return glm::lookAt(position, position + forward, up);
+}
+
+inline glm::mat4 Camera::GetProjectionMatrix() const noexcept
+{
+	const float height		= viewport.height;
+	const float width		= viewport.width;
+	const float aspectRatio = width / height;
+
+	switch (projection)
+	{
+		case Projection::Orthographic:
+		{
+			const float halfHeight = orthoSize * 0.5f;
+			const float halfWidth  = halfHeight * aspectRatio;
+
+			return glm::ortho(-halfWidth, halfWidth, -halfHeight, halfHeight, nearPlane, farPlane);
+		}
+		case Projection::Perspective:
+		{
+			return glm::perspective(glm::radians(fov), aspectRatio, nearPlane, farPlane);
+		}
+		default:
+		{
+			return { 1.0f };
+		}
+	}
+}
