@@ -1,32 +1,23 @@
-#ifndef GUARD_MARBLE_H
+﻿#ifndef GUARD_MARBLE_H
 #define GUARD_MARBLE_H
 
 #include "Object.h"
 
-/**
- * @brief ����.
- */
+class Mesh;
+
 class Marble final
 	: public Object
 {
 public:
 	/**
-	 * @brief ������.
+	 * @brief 
 	 * 
-	 * @param boundary_ ������ ������ ������ �� �ִ� ��谪
-	 * @param position_ ������ ������ ��ġ
-	 * @param rotation_ ������ ������ ȸ��
-	 * @param scale_	������ ������ ũ��
+	 * @param mesh_ 
 	 */
-	explicit Marble(
-		const float      boundary_,
-		const glm::vec3& position_ = glm::vec3(0.0f),
-		const glm::vec3& rotation_ = glm::vec3(0.0f),
-		const glm::vec3& scale_    = glm::vec3(1.0f)
-	) noexcept;
+	explicit Marble(Mesh* const mesh_) noexcept;
 
 	/**
-	 * @brief �Ҹ���.
+	 * @brief 
 	 */
 	virtual ~Marble() noexcept override;
 
@@ -37,21 +28,23 @@ public:
 	 */
 	virtual void Update(const float deltaTime_) noexcept override;
 
+	/**
+	 * @brief 
+	 * 
+	 * @param shader_ 
+	 */
+	virtual void Render(const Shader& shader_) const noexcept override;
+
 private:
 	/**
-	 * @brief �ش� ������ �̵� �ӵ�.
+	 * @brief 해당 구슬의 매쉬.
 	 */
-	static constexpr float MOVE_SPEED = 10;
+	Mesh* mesh;
 
-	/**
-	 * @brief ��谪
-	 */
-	float boundary;
-
-	/**
-	 * @brief �ش� ������ �̵� ����.
-	 */
-	glm::vec3 direction;
+	glm::vec3  velocity = glm::vec3(3.0f, 2.0f, 0.0f); // 초기 속도
+	float      radius = 0.5f;       // 메쉬 크기에 맞게 조정
+	float      restitution = 0.85f;      // 반발계수
+	float      damping = 0.995f;     // 미세 감쇠
 };
 
 #endif // !GUARD_MARBLE_H
