@@ -1,95 +1,144 @@
-#ifndef GUARD_MESH_H
-#define GUARD_MESH_H
+﻿#pragma once
 
-#include <string>
+#include "PCH.h"
 
-#include <GL/glew.h>
-
-#include <glm/vec3.hpp>
-
-/**
- * @brief �޽��� �����մϴ�.
- */
 class Mesh
 {
 public:
 	/**
-	 * @brief ������ �����մϴ�.
+	 * @brief 정점 데이터를 정의합니다.
 	 */
 	struct Vertex final
-	{ 
+	{
 		/**
-		 * @brief ���� ��ġ.
+		 * @brief 정점의 위치.
 		 */
 		glm::vec3 position;
 
 		/**
-		 * @brief ���� ��.
+		 * @brief 정점의 색상.
 		 */
 		glm::vec3 color;
-	};
-
+	};    
+    
     /**
-     * @brief ������.
+     * @brief 생성자.
      *
-     * @param vertices_ ������ ������.
-     * @param indices_  ������ �ε�����.
+     * @param vertices_ 생성할 정점들.
+     * @param indices_  생성할 인덱스들.
      */
-    explicit Mesh(
-        const std::vector<Vertex>& vertices_,
-        const std::vector<GLuint>& indices_
-    ) noexcept;
+    explicit Mesh(const std::vector<Vertex>& vertices_,
+                  const std::vector<GLuint>& indices_) noexcept;
 
     /**
-     * @brief �Ҹ���.
+     * @brief 소멸자.
      */
     virtual ~Mesh();
 
     /**
-     * @brief �ش� �޽��� �������մϴ�.
-     */
-    void Render() const noexcept;
+     * @brief 해당 메쉬의 정점들을 반환합니다.
+     *
+     * @return std::vector<Vertex>& 해당 메쉬의 정점들.
+	 */
+    [[nodiscard]]
+    inline constexpr const std::vector<Vertex>& GetVertices() const noexcept;
 
     /**
-     * @brief ���Ͽ��� �޽��� �ε��մϴ�.
+     * @brief 해당 메쉬의 정점들을 설정합니다.
      *
-     * @param filePath_ �ε��� ���� ���.
+	 * @param vertices_ 설정할 정점들.
+     */
+	inline void SetVertices(const std::vector<Vertex>& vertices_) noexcept;
+
+    /**
+     * @brief 해당 메쉬의 인덱스들을 반환합니다.
      *
-     * @return Mesh* �ε�� �޽� ��ü.
+     * @return std::vector<GLuint>& 해당 메쉬의 인덱스들.
+	 */
+    [[nodiscard]]
+	inline constexpr const std::vector<GLuint>& GetIndices() const noexcept;
+
+    /**
+     * @brief 해당 메쉬의 인덱스들을 설정합니다.
+	 *
+     */
+	inline void SetIndices(const std::vector<GLuint>& indices_) noexcept;
+
+    /**
+     * @brief 해당 메쉬의 정점 데이터를 업데이트합니다.
+	 */
+	void UpdateVertices() const noexcept;
+
+    /**
+     * @brief 해당 메쉬의 인덱스 데이터를 업데이트합니다.
+	 */
+	void UpdateIndices() const noexcept;
+
+    /**
+     * @brief 해당 메쉬를 렌더링합니다.
+     *
+     * @param renderMode_ 렌더링 모드. 기본값은 GL_TRIANGLES입니다.
+     */
+    void Render(GLenum renderMode_ = GL_TRIANGLES) const noexcept;
+
+    /**
+     * @brief 파일에서 메쉬를 로드합니다.
+     *
+     * @param filePath_ 로드할 파일 경로.
+     *
+     * @return Mesh* 로드된 메쉬 객체.
      */
     [[nodiscard]]
-    static Mesh* LoadFrom(const std::string& filePath_) noexcept;
+    static std::unique_ptr<Mesh> LoadFrom(const std::string& filePath_) noexcept;
 
 private:
     /**
-     * @brief �ش� ���� �迭 ��ü.
+     * @brief 해당 정점 배열 객체.
      */
     unsigned int vao;
 
     /**
-     * @brief �ش� �Ž��� ���� ���� ��ü.
+     * @brief 해당 매쉬의 정점 버퍼 객체.
      */
     unsigned int vbo;
 
     /**
-     * @brief �ش� �Ž��� ��� �迭 ��ü.
+     * @brief 해당 매쉬의 요소 배열 객체.
      */
     unsigned int ebo;
 
     /**
-     * @brief �ش� �Ž��� ���� �����͵�.
+     * @brief 해당 매쉬의 정점 데이터들.
      */
     std::vector<Vertex> vertices;
 
     /**
-     * @brief �ش� �Ž��� �ε�����.
+     * @brief 해당 매쉬의 인덱스들.
      */
     std::vector<GLuint> indices;
 
     /**
-     * @brief �ش� �޽��� �ʱ�ȭ ����.
+     * @brief 해당 메쉬의 초기화 여부.
      */
     bool isInitialized;
 };
 
-#endif // !GUARD_MESH_H
+inline constexpr const std::vector<Mesh::Vertex>& Mesh::GetVertices() const noexcept
+{
+    return vertices;
+}
+
+inline void Mesh::SetVertices(const std::vector<Vertex>& vertices_) noexcept
+{
+    vertices = vertices_;
+}
+
+inline constexpr const std::vector<GLuint>& Mesh::GetIndices() const noexcept
+{
+    return indices;
+}
+
+inline void Mesh::SetIndices(const std::vector<GLuint>& indices_) noexcept
+{
+    indices = indices_;
+}
