@@ -1,196 +1,177 @@
-#ifndef GUARD_OBJECTS_H
-#define GUARD_OBJECTS_H
+﻿#pragma once
 
-#include <memory>
+#include "PCH.h"
 
-#include <glm/vec3.hpp>
-
-#include "Mesh.h"
 #include "Shader.h"
 
-struct Transform final
-{
-    glm::vec3 position = glm::vec3(0 , 0, 0);
-    glm::vec3 rotation = glm::vec3(0, 0, 0);
-	glm::vec3 scale = glm::vec3(1, 1, 1);
-    Transform* parent = nullptr;
-
-    inline const glm::mat4 GetModelMatrix() const noexcept
-    {
-		glm::mat4 model = glm::mat4(1.0f);
-		model = glm::translate(model, position);
-		model = glm::rotate(model, glm::radians(rotation.x), glm::vec3(1.0f, 0.0f, 0.0f));
-		model = glm::rotate(model, glm::radians(rotation.y), glm::vec3(0.0f, 1.0f, 0.0f));
-		model = glm::rotate(model, glm::radians(rotation.z), glm::vec3(0.0f, 0.0f, 1.0f));
-		model = glm::scale(model, scale);
-
-        if (parent)
-        {
-            model = parent->GetModelMatrix() * model;
-        }
-
-        return model;
-	}
-};
-
 /**
- * @class Object
- *
- * @brief ������Ʈ�� �����մϴ�.
+ * @brief 
  */
 class Object
 {
 public:
-    /**
-     * @brief ������.
-     *
-     * @param position_ ������ ������Ʈ�� ��ġ
-     * @param rotation_ ������ ������Ʈ�� ȸ��
-     * @param scale_    ������ ������Ʈ�� ũ��
-     */
-    explicit Object(
-        const glm::vec3& position_ = glm::vec3{ 0.0f, 0.0f, 0.0f },
-        const glm::vec3& rotation_ = glm::vec3{ 0.0f, 0.0f, 0.0f },
-        const glm::vec3& scale_    = glm::vec3{ 1.0f, 1.0f, 1.0f }
-    ) noexcept;
+	/**
+	 * @brief 소멸자.
+	 */
+	virtual ~Object() noexcept = default;
+
+	/**
+	 * @brief 해당 객체를 업데이트합니다.
+     * 
+	 * @param deltaTime_ 이전 프레임과 현재 프레임 사이의 간격
+	 */
+	virtual void Update(const float deltaTime_) noexcept;
 
     /**
-     * @brief �Ҹ���.
-     */
-    virtual ~Object() noexcept;
-
-    /**
-     * @brief �ش� ������Ʈ�� ������Ʈ�մϴ�.
-     */
-    virtual void Update(const float deltaTime_) noexcept;
-
-    /**
-     * @brief �ش� ������Ʈ�� �������մϴ�.
-     *
-     * @param shader_ ����� ���̴�
+     * @brief 해당 객체를 렌더링합니다.
+     * 
+     * @param shader_ 사용할 셰이더
      */
     virtual void Render(const Shader& shader_) const noexcept;
 
     /**
-     * @brief �ش� ������Ʈ�� �Ž��� ��ȯ�մϴ�.
+     * @brief 해당 오브젝트의 위치를 반환합니다.
      *
-     * @return Mesh* �ش� ������Ʈ�� �Ž�
-     */
-    [[nodiscard]]
-    inline Mesh* GetMesh() const noexcept;
-
-    /**
-     * @brief �ش� ������Ʈ�� �Ž��� �����մϴ�.
-     *
-     * @param mesh_ ������ �Ž�.
-     */
-    inline void SetMesh(Mesh* mesh_) noexcept;
-
-    /**
-     * @brief �ش� ������Ʈ�� ��ġ�� ��ȯ�մϴ�.
-     *
-     * @return glm::vec3 �ش� ������Ʈ�� ��ġ.
+     * @return glm::vec3 해당 오브젝트의 위치
      */
     [[nodiscard]]
     inline constexpr glm::vec3 GetPosition() const noexcept;
 
     /**
-     * @brief �ش� ������Ʈ�� ��ġ�� �����մϴ�.
+     * @brief 해당 오브젝트의 위치를 설정합니다.
      *
-     * @param position_ ������ ��ġ.
+     * @param position_ 설정할 위치
      */
     inline void SetPosition(const glm::vec3& position_) noexcept;
 
     /**
-     * @brief �ش� ������Ʈ�� ȸ���� ��ȯ�մϴ�.
+     * @brief 해당 오브젝트의 회전을 반환합니다.
      *
-     * @return glm::vec3 �ش� ������Ʈ�� ȸ��.
+     * @return glm::vec3 해당 오브젝트의 회전
      */
     [[nodiscard]]
     inline constexpr glm::vec3 GetRotation() const noexcept;
 
     /**
-     * @brief �ش� ������Ʈ�� ȸ���� �����մϴ�.
+     * @brief 해당 오브젝트의 회전을 설정합니다.
      *
-     * @param rotation_ ������ ȸ��.
+     * @param rotation_ 설정할 회전
      */
     inline void SetRotation(const glm::vec3& rotation_) noexcept;
 
     /**
-     * @brief �ش� ������Ʈ�� ũ�⸦ ��ȯ�մϴ�.
+     * @brief 해당 오브젝트의 크기를 반환합니다.
      *
-     * @return glm::vec3 �ش� ������Ʈ�� ũ��.
+     * @return glm::vec3 해당 오브젝트의 크기
      */
     [[nodiscard]]
     inline constexpr glm::vec3 GetScale() const noexcept;
 
     /**
-     * @brief �ش� ������Ʈ�� ũ�⸦ �����մϴ�.
+     * @brief 해당 오브젝트의 크기를 설정합니다.
      *
-     * @param scale_ ������ ũ��.
+     * @param scale_ 설정할 크기
      */
     inline void SetScale(const glm::vec3& scale_) noexcept;
 
-    inline const Transform& GetTransform() const noexcept
-    {
-        return transform;
-	}
+    /**
+     * @brief 해당 오브젝트의 크기를 반환합니다.
+     *
+     * @return glm::vec3 해당 오브젝트의 크기
+     */
+    [[nodiscard]]
+    inline constexpr Object* const GetParent() const noexcept;
 
-    void SetParent(Object* parent_) noexcept
-    {
-        transform.parent = parent_ ? &parent_->transform : nullptr;
-    }
+    /**
+     * @brief 해당 오브젝트의 크기를 설정합니다.
+     *
+     * @param scale_ 설정할 크기
+     */
+    inline void SetParent(Object* const parent_) noexcept;
+
+    /**
+     * @brief 해당 오브젝트의 모델 행렬을 반환합니다.
+     * 
+	 * @return glm::mat4 해당 오브젝트의 모델 행렬
+     */
+    [[nodiscard]]
+	inline glm::mat4 GetModelMatrix() const noexcept;
 
 private:
-    /**
-     * @brief �ش� ������Ʈ�� �Ž�.
-     */
-    std::unique_ptr<Mesh> mesh;
-
-    /**
-     * @brief �ش� ������Ʈ�� ��ġ.
+	/**
+	 * @brief 해당 객체의 위치.
 	 */
-    Transform transform;
+	glm::vec3 position = glm::vec3(0.0f);
+
+	/**
+	 * @brief 해당 객체의 회전.
+	 */
+    glm::vec3 rotation = glm::vec3(0.0f);
+
+	/**
+	 * @brief 해당 객체의 크기.
+	 */
+	glm::vec3 scale = glm::vec3(1.0f);
+
+	/**
+	 * @brief 해당 객체의 상위 객체.
+	 */
+	Object* parent = nullptr;
 };
-
-inline Mesh* Object::GetMesh() const noexcept
-{
-    return mesh.get();
-}
-
-inline void Object::SetMesh(Mesh* const mesh_) noexcept
-{
-    mesh.reset(mesh_);
-}
 
 inline constexpr glm::vec3 Object::GetPosition() const noexcept
 {
-    return transform.position;
+    return position;
 }
 
 inline void Object::SetPosition(const glm::vec3& position_) noexcept
 {
-    transform.position = position_;
+    position = position_;
 }
 
 inline constexpr glm::vec3 Object::GetRotation() const noexcept
 {
-    return transform.rotation;
+    return rotation;
 }
 
 inline void Object::SetRotation(const glm::vec3& rotation_) noexcept
 {
-    transform.rotation = rotation_;
+    rotation = rotation_;
 }
 
 inline constexpr glm::vec3 Object::GetScale() const noexcept
 {
-    return transform.scale;
+    return scale;
 }
 
 inline void Object::SetScale(const glm::vec3& scale_) noexcept
 {
-    transform.scale = scale_;
+    scale = scale_;
 }
 
-#endif // !GUARD_OBJECTS_H
+inline constexpr Object* const Object::GetParent() const noexcept
+{
+    return parent;
+}
+
+inline void Object::SetParent(Object* const parent_) noexcept
+{
+    parent = parent_;
+}
+
+inline glm::mat4 Object::GetModelMatrix() const noexcept
+{
+    glm::mat4 model = glm::mat4(1.0f);
+    model = glm::translate(model, position);
+    model = glm::rotate(model, glm::radians(rotation.x), glm::vec3(1.0f, 0.0f, 0.0f));
+    model = glm::rotate(model, glm::radians(rotation.y), glm::vec3(0.0f, 1.0f, 0.0f));
+    model = glm::rotate(model, glm::radians(rotation.z), glm::vec3(0.0f, 0.0f, 1.0f));
+    model = glm::scale(model, scale);
+
+    if (parent != nullptr)
+    {
+        model = parent->GetModelMatrix() * model;
+    }
+
+    return model;
+}
