@@ -1,123 +1,95 @@
-#ifndef GUARD_SHADER_H
-#define GUARD_SHADER_H
+﻿#pragma once
 
-#include <filesystem>
-#include <memory>
-
-#include <gl/glew.h>
-
-#include <glm/vec2.hpp>
-#include <glm/vec3.hpp>
-#include <glm/vec4.hpp>
-#include <glm/mat4x4.hpp>
-#include <glm/gtc/type_ptr.hpp>
+#include "PCH.h"
 
 /**
  * @class Shader
  *
- * @brief ���̴��� �����մϴ�.
+ * @brief 셰이더를 정의합니다.
  */
 class Shader final
 {
 public:
     /**
-     * @brief ������.
-     *
-     * @param vertexSource_   ���ؽ� ���̴� �ҽ� �ڵ�.
-     * @param fragmentSource_ �����׸�Ʈ ���̴� �ҽ� �ڵ�
+     * @brief 생성자.
      */
-    explicit Shader(
-        const char* vertexSource_,
-        const char* fragmentSource_
-    ) noexcept;
+    explicit Shader() noexcept;
 
     /**
-     * @brief �Ҹ���.
+     * @brief 소멸자.
      */
     ~Shader() noexcept;
 
     /**
-     * @brief ���̴��� ����մϴ�.
+     * @brief 셰이더를 사용합니다.
      */
     inline void Use() const noexcept;
 
     /**
-     * @brief ���̴� ���α׷� ID�� ��ȯ�մϴ�.
+     * @brief 셰이더 프로그램 ID를 반환합니다.
      *
-     * @return unsigned int ���̴� ���α׷� ID.
+     * @return unsigned int 셰이더 프로그램 ID.
      */
     [[nodiscard]]
     constexpr unsigned int GetProgramID() const noexcept;
 
     /**
-     * @brief ������ ������ ������ �����մϴ�.
+     * @brief 정수형 유니폼 변수를 설정합니다.
      *
-     * @param name_  ������ ���� �̸�.
-     * @param value_ ������ ��.
+     * @param name_  유니폼 변수 이름.
+     * @param value_ 설정할 값.
      */
-    inline void SetUniformInt(
-        const char* const name_, 
-        const int         value_
-    ) const noexcept;
+    inline void SetUniformInt(const char* const name_,
+                              const int         value_) const noexcept;
 
     /**
-     * @brief �Ǽ��� ������ ������ �����մϴ�.
+     * @brief 실수형 유니폼 변수를 설정합니다.
      *
-     * @param name_  ������ ���� �̸�.
-     * @param value_ ������ ��.
+     * @param name_  유니폼 변수 이름.
+     * @param value_ 설정할 값.
      */
-    inline void SetUniformFloat(
-        const char* const name_, 
-        const float       value_
-    ) const noexcept;
+    inline void SetUniformFloat(const char* const name_,
+                                const float       value_) const noexcept;
 
     /**
-     * @brief 2���� ���� ������ ������ �����մϴ�.
+     * @brief 2차원 벡터 유니폼 변수를 설정합니다.
      *
-     * @param name_  ������ ���� �̸�.
-     * @param value_ ������ ��.
+     * @param name_  유니폼 변수 이름.
+     * @param value_ 설정할 값.
      */
-    inline void SetUniformVector2(
-        const char* const name_, 
-        const glm::vec2&  value_
-    ) const noexcept;
+    inline void SetUniformVector2(const char* const name_,
+                                  const glm::vec2& value_) const noexcept;
 
     /**
-     * @brief 3���� ���� ������ ������ �����մϴ�.
+     * @brief 3차원 벡터 유니폼 변수를 설정합니다.
      *
-     * @param name_  ������ ���� �̸�.
-     * @param value_ ������ ��.
+     * @param name_  유니폼 변수 이름.
+     * @param value_ 설정할 값.
      */
-    inline void SetUniformVector3(
-        const char* const name_, 
-        const glm::vec3&  value_
-    ) const noexcept;
+    inline void SetUniformVector3(const char* const name_,
+                                  const glm::vec3& value_) const noexcept;
 
     /**
-     * @brief 4���� ���� ������ ������ �����մϴ�.
+     * @brief 4차원 벡터 유니폼 변수를 설정합니다.
      *
-     * @param name_  ������ ���� �̸�.
-     * @param value_ ������ ��.
+     * @param name_  유니폼 변수 이름.
+     * @param value_ 설정할 값.
      */
-    inline void SetUniformVector4(
-        const char* const name_, 
-        const glm::vec4&  value_
-    ) const noexcept;
+    inline void SetUniformVector4(const char* const name_,
+                                  const glm::vec4& value_) const noexcept;
 
     /**
-     * @brief 4x4 ��� ������ ������ �����մϴ�.
+     * @brief 4x4 행렬 유니폼 변수를 설정합니다.
      *
-     * @param name_  ������ ���� �̸�.
-     * @param value_ ������ ��.
+     * @param name_  유니폼 변수 이름.
+     * @param value_ 설정할 값.
      */
-    inline void SetUniformMatrix4x4(
-        const char* const name_, 
-        const glm::mat4&  value_
-    ) const noexcept;
+    inline void SetUniformMatrix4x4(const char* const name_,
+                                    const glm::mat4& value_) const noexcept;
 
 private:
     /**
-     * @brief ���̴� ���α׷� ID.
+     * @brief 셰이더 프로그램 ID.
      */
     GLuint programID = 0;
 };
@@ -167,5 +139,3 @@ inline void Shader::SetUniformMatrix4x4(const char* name_, const glm::mat4& valu
     const GLint location = glGetUniformLocation(programID, name_);
     glUniformMatrix4fv(location, 1, GL_FALSE, glm::value_ptr(value_));
 }
-
-#endif // !GUARD_SHADER_H
